requirements.txt
paho-mqtt
numpy
PyMySQL
python-dotenv
